Add comment feature!
Now you can comment on posts! There's a nice neat little comment
form that shows below a post (only iff you're logged in), and you
can use it to post a comment. Which is neato.

The comment controller only has the method for CREATE at this point.
So yeah, there's no going back.