Bootstrap it!
This thing was dog-ugly before. Now it's working towards being bootstrappy.
I also moved a lot of repeated code into partials.

There are still missing links and buttons and stuff that's out of place,
but this is a lot prettier.

We now include all the bootstrap stuff in our asset pipeline, also, so that's
cool.